Move Double+Stats to WordPressShared
